Add some javadocs Refactored files into new package
