+++ src/main/java/switch_managers/ManagerController.java
@@ -1,8 +1,11 @@
-package interfaces;
+package switch_managers;
 
 import core.StateManager;
-import core.SwitchEventType;
 
+/**
+ * Manager Controller Interface
+ * Used to handle switches between managers.
+ */
 public interface ManagerController {
     public StateManager switchManagers(SwitchEventType type, StateManager currManager);
 }
