+++ src/main/java/switch_managers/SwitchEventType.java
@@ -1,4 +1,4 @@
-package core;
+package switch_managers;
 
 public enum SwitchEventType {
     PAUSE,
